requirements.txt
pymongo
python-dotenv
beautifulsoup4
nltk
soupsieve
click
joblib
regex
tqdm
flask
requests